fixed voting button style and functionality
