+++ styles/globals.css
@@ -37,6 +37,11 @@ button {
     "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif !important;
 }
 
+button.font-normal {
+  font-family: "Open Sans", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell,
+    "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+}
+
 * {
   box-sizing: border-box;
 }
